refactor: move get-sidechain-params command logic to a library crate
