+++ primitives/sidechain/Cargo.toml
@@ -12,6 +12,8 @@ parity-scale-codec = { workspace = true }
 scale-info = { workspace = true }
 sp-api = { workspace = true }
 sp-runtime = { workspace = true }
+sp-blockchain = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
 
 [features]
 default = ["std"]
@@ -22,5 +24,7 @@ std = [
     "scale-info/std",
 	"sidechain-domain/std",
     "sp-api/std",
-    "sp-runtime/std"
+    "sp-runtime/std",
+	"serde_json",
+	"sp-blockchain"
 ]
